quiz/quiz/requirements.txt
Django==3.0.5
django-nested-admin==3.3.0